news/requirements.txt
beautifulsoup4==4.11.1
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.20.0
Markdown==3.3.7
requests==2.28.0
urllib3==1.26.9
celery==5.2.7
redis==4.3.3
djangorestframework-simplejwt==5.2.0
pytest==7.1.2
pytest-django==4.5.2